fix: unpin after all the data is consumed & fix tests for server

+++ storage-node/src/cache/data_store_cache/memdisk/data_store/disk.rs
@@ -1,11 +1,17 @@
-use std::fs;
+use std::{fs, sync::Arc};
 
 use bytes::Bytes;
 use futures::StreamExt;
-use tokio::sync::mpsc::Receiver;
+use tokio::sync::{mpsc::Receiver, Mutex};
 
 use crate::{
-    disk::disk_manager::DiskManager, error::ParpulseResult, storage_reader::StorageReaderStream,
+    cache::{
+        data_store_cache::memdisk::{MemDiskStoreReplacerKey, MemDiskStoreReplacerValue},
+        replacer::DataStoreReplacer,
+    },
+    disk::disk_manager::DiskManager,
+    error::ParpulseResult,
+    storage_reader::StorageReaderStream,
 };
 
 /// TODO(lanlou): make them configurable.
@@ -22,7 +28,10 @@ pub struct DiskStore {
 
 impl Drop for DiskStore {
     fn drop(&mut self) {
-        fs::remove_dir_all(self.base_path.clone()).expect("remove cache files failed");
+        if fs::metadata(&self.base_path).is_ok() {
+            println!("Removing cache files: {:?}", self.base_path);
+            fs::remove_dir_all(self.base_path.clone()).expect("remove cache files failed");
+        }
     }
 }
 
@@ -32,7 +41,6 @@ impl DiskStore {
         if !final_base_path.ends_with('/') {
             final_base_path += "/";
         }
-
         Self {
             disk_manager,
             base_path: final_base_path,
@@ -43,10 +51,14 @@ impl DiskStore {
 impl DiskStore {
     /// Reads data from the disk store. The method returns a stream of data read from the disk
     /// store.
-    pub async fn read_data(
+    pub async fn read_data<R>(
         &self,
         key: &str,
-    ) -> ParpulseResult<Option<Receiver<ParpulseResult<Bytes>>>> {
+        disk_replacer: Arc<Mutex<R>>,
+    ) -> ParpulseResult<Option<Receiver<ParpulseResult<Bytes>>>>
+    where
+        R: DataStoreReplacer<MemDiskStoreReplacerKey, MemDiskStoreReplacerValue> + 'static,
+    {
         // TODO(lanlou): we later may consider the remaining space to decide the buffer size
         let mut buffer_size = self.disk_manager.file_size(key).await? as usize;
         if buffer_size > MAX_DISK_READER_BUFFER_SIZE {
@@ -55,6 +67,7 @@ impl DiskStore {
         // FIXME: Shall we consider the situation where the data is not found?
         let mut disk_stream = self.disk_manager.disk_read_stream(key, buffer_size).await?;
         let (tx, rx) = tokio::sync::mpsc::channel(DEFAULT_DISK_CHANNEL_BUFFER_SIZE);
+        let key_str = key.to_string().clone();
         tokio::spawn(async move {
             loop {
                 match disk_stream.next().await {
@@ -67,6 +80,8 @@ impl DiskStore {
                     None => break,
                 }
             }
+            // TODO(lanlou): when second read, so there is no need to unpin, how to improve?
+            disk_replacer.lock().await.unpin(&key_str);
         });
         Ok(Some(rx))
     }
